services/api-gateway/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx>=0.25.0
slowapi==0.1.9
python-json-logger==2.0.7

